lots of spelling errors (refactoring)
