+++ RepetierHost/RepetierHost/windows/Preferences.xib
@@ -2288,7 +2288,7 @@
 																								<object class="NSTextFieldCell" key="NSCell" id="724110757">
 																									<int key="NSCellFlags">68157504</int>
 																									<int key="NSCellFlags2">272630784</int>
-																									<string key="NSContents">Visualitazion Quality:</string>
+																									<string key="NSContents">Visualization Quality:</string>
 																									<reference key="NSSupport" ref="27"/>
 																									<string key="NSCellIdentifier">_NS:3944</string>
 																									<reference key="NSControlView" ref="820096751"/>
